把增加供应商的上传文件功能给加上了

+++ psi-cpp/psi-c1/domain/query/supplierQuery/SpecifiedSupplierDataQuery.h
@@ -8,7 +8,7 @@
 class SpecifiedSupplierDataQuery 
 {
 	// ±àºÅ
-	CC_SYNTHESIZE(int, code, Code);
+	CC_SYNTHESIZE(uint64_t, code, Code);
 	
 public:
 	// °ó¶¨from_json
